+++ apps/common-app/runtime-tests/reanimated/suites.ts
@@ -68,6 +68,7 @@ export const REANIMATED_TEST_SUITES: RuntimeTestSuite[] = [
   {
     testSuiteName: 'props',
     importTest: () => {
+      require('./tests/props/backgroundImage.test');
       require('./tests/props/boxShadow.test');
     },
   },