+++ packages/react-native-reanimated/src/hook/useAnimatedReaction.ts
@@ -33,7 +33,7 @@ export function useAnimatedReaction<PreparedResult>(
   >,
   dependencies?: DependencyList
 ) {
-  let inputs = Object.values(prepare.__closure ?? {});
+  let inputs = prepare.__closure ?? [];
 
   if (!inputs.length && dependencies?.length) {
     // let web work without Worklets Babel plugin